Updates Types example to parse recursive types.
Previously, the types example did not support parsing of recursive
types, despite the fact that the underlying system does support them.
Therefore, I've made a small modification to the parser to allow them to
be entered.  The syntax is "\X.T" where X is the variable name, and T
the type which can contain a reference to X.

Unfortunately, at this time, there is still a problem with printing
recursive automata which is unknown.