Allows sorting activations using a comparator #13
The comparator is used to sort activations for each state which it is
encountered.  A default "RankComparator" is used for this.  The sorting
does have a clear effect on the number of rewrites required.  However,
it's not clear at the moment whether the default comparator has a
positive or negative effect.