Added concept of "RewritingNormaliser" #17
Have replaced the slightly odd concept of a "stacked rewrite" with that
or a "RewritingNormaliser" and the general concept of a normaliser.
Thus, after a successful rule application is made the normaliser is run
before the resulting automaton is registered in the rewrite.  This means
the normaliser can run minimisation and/or garbage collection, etc.