Update Rewriter API #17
This is a fairly significant update to the Rewriter API.  The main
change lies with the way that inference rules are implemented, compared
with rewrite rules.  However, currently, this change has led to some
reasonable degradation in performance and future improvements will be
needed to address this.  The structure and organisation of the Rewriter
API has also been changed, and I think is now more coherent that before.

Inference rules were previously using rewrite, like reductions.
However, instead of this, I've updated them to use substitute.  This
means they do not throw away information, as they did before.  The
downside, however, is that automaton tend to grow much larger than
before.  This means that basic automaton operations (e.g. adding new
states, minimising, compacting, etc) become more expensive since their
cost is currently related to automaton size.  In principle, this
dependency could be improved or remove altogether which could lead to
useful performance improvements.

One major change is that inference rules must now have a specific root
on which they are operating.  This is necessary so that they can
substitute from this root.  Therefore, rewrite rules and activations are
now separate from inference rules and activations, with the latter
including a concept of root during the rewrite itself.