Update to support both inference modes
This update is something of an odd one.  In essence, what I've done is
enabled two modes of operation.  In one mode, inferences are handled
using substitution and in the other they are handled using rewrites.
The reason for this is that the rewrite-based approach remains currently
the fastest approach.  However, I believe that it is also a flawed
approach and that I should be using the substitution approach.
Therefore, I will support both modes until such time that I can properly
retire the rewrite-based approach.