Update incremental minimiser #19
The incremental minimiser is about ready to be tested.  It is definitely
not working as efficiently as I would like at this stage, but I have
maximised the reuse of existing automata code.  This is just to increase
the likelyhood of success!

To begin testing it, I need to plumb it into the Reducer somehow so that
Reduction rules no longer call "minimise", and that the incremental
minimiser is called instead.