Implement incremental compaction
The incremental compaction algorithm is in place, but unfortuantely
cannot be used.  The reason for this is that, unfortunately, the
substitute() function which can be called from the body of a rewrite
rule activation calls minimise() which leaves the automaton
with empty states.  Thus, unfortunately, we have to defer to an
inefficient algorithm which scans all states.