Implement Bipartite Matching #31
During automaton minimisation, the way that bag states were being
compared for equivalence was plain wrong.  In fact, it turns out the
problem is equivalent to that of determining whether a perfect matching
exists for a bipartite graph.  In this case, I've implemented a variant
of "Kuhn's" algorithm which does this pretty efficiently.